src/requirements.txt
Flask>=2.0
boto3>=1.20
python-dotenv>=0.19
